% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.pc.R
\name{assoc.vars}
\alias{assoc.vars}
\title{Biostats Principal Component Analysis}
\usage{
assoc.vars(d, pd, nv, fn, vadd = NULL)
}
\arguments{
\item{pd}{A dataframe containing sample metadata.}

\item{nv}{A character vector of covariates to test.}

\item{pc}{A prcomp object.}
}
\description{
Tests for correlations between technical covariates and PCA dimensions.
}
