% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roasthtml.R
\name{write.html.mod2}
\alias{write.html.mod2}
\title{Core HTML Assembler and Dynamic Table Generation Wrapper}
\usage{
write.html.mod2(
  x,
  file = paste0(htmlpath, htmlname),
  links = links,
  tiny.pic = plots,
  sorttable = sorttable,
  dragtable = dragtable,
  css = "../../../../../biostats.css",
  title = title
)
}
\arguments{
\item{x}{A structured data.frame to print.}

\item{file}{Character. Full path configuration target for file generation.}

\item{links}{A structured list containing column-matched URL targets.}

\item{tiny.pic}{A structured list containing column-matched graphic paths.}

\item{sorttable}{Character string containing contents of `sorttable.js`.}

\item{dragtable}{Character string containing contents of `dragtable.js`.}

\item{css}{Character. Relative path configuration target for CSS files.}

\item{title}{Character. Text element to serve as the document title.}
}
\description{
Writes dynamic table elements, formats embedded hyperlinks/images, and
outputs unified files utilizing internal `hwriter` constructs.
}
