% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roasthtml.R
\name{roastHtmlDETable}
\alias{roastHtmlDETable}
\title{Export Nested Differential Expression Target Layout Sub-tables}
\usage{
roastHtmlDETable(
  mygs,
  gs,
  i,
  detable,
  DEdir,
  filename,
  dirname,
  maxgs = 50,
  verbose = TRUE,
  selcols,
  vorder
)
}
\arguments{
\item{mygs}{A specific \code{roastgsa} results structure object.}

\item{gs}{Character. The target name context matching background collections.}

\item{i}{Character. The specific contrast assignment tracker.}

\item{detable}{A data.frame mapping baseline differential parameters.}

\item{DEdir}{Character. Input file storage directory referencing static plots.}

\item{filename}{Character. Output name specification tracking target templates.}

\item{dirname}{Character. Container folder path destination mapping components.}

\item{maxgs}{Integer. Maximum scale tracking threshold. Default 50.}

\item{verbose}{Logical. Enables output standard messages. Default TRUE.}

\item{selcols}{Vector of characters filtering acceptable column outputs.}

\item{vorder}{Character string declaring variable order hierarchy sequence targeting.}
}
\description{
Evaluates specific matching features between target gene set lists and dynamic sequence structures,
writing localized independent HTML tracking sheets mapping data frames into interactive sortable models.
}
